--------
A      supernode.c

--------

A: Added code to send REGISTER_ACK on receipt of REGISTER.


Commentary

This should always have been done but did not yet cause problems. I am doing
this now as this behaviour is required when multiple supernodes exist and the
edge must decide which are working. The multiple supernode feature is intended
to be added soon.

In addition to this, have REGISTER_ACK sent back allows the edge to retry
registration when it doesn't get REGISTER_ACK after some timeout.


git-svn-id: https://svn.ntop.org/svn/ntop/trunk/n2n@3487 a02cd0c1-8e76-42e1-a119-56f9641475e2